third_party/blink/web_tests/platform/win11-arm64/virtual/fenced-frame-mparch/external/wpt/html/anonymous-iframe/embedding.tentative.https.window_4-4-expected.txt

This is a testharness.js-based test.
FAIL COEP:require-corp parent embeds cross-origin credentialless iframe assert_equals: expected "load" but got "block"
Harness: the test ran to completion.
